brick: Pass the index of the source in the destination edges array.
Thanks to this patch the destination will be able to know which edge in it's
arrays point to the source of the packet.

This will be used by component who associate the source of a packet with it's
content. For example an ethernet switch associating the source with the source
mac address.

Signed-off-by: Benoît Canet <benoit.canet@nodalink.com>