brick: Add a nic brick
This brick is used to communicate with a ethernet, or a pcap port
The reading part is implemented in the pull callback and the writing
part is in the burst callback.

This brick is configured by the name of the interface or it's id and by the
output side to use.

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>
Signed-off-by: Jérôme Jutteau <jerome.jutteau@outscale.com>