vhost-user: Properly kill and wait for QEMU exist in the tests
The glib documentation specifies that g_spawn_close_pid is doing
nothing under UNIX so the code need to properly kill QEMU and
wait it exit.

This bug was causing the vhost-user test to randomly hang since a
previous QEMU instance had not been terminated and was
using all the huge pages.

Signed-off-by: Benoît Canet <benoit.canet@nodalink.com>