mac-table: add mac-table
Add a new file mac-table which contain some function used to manipulate
an array where a pg_mac is use as the key of this array.
This array can be use to store "void *" or "elements"
An elements is an undefine type, the size of this type need to be set
setting a new elements.
All the elements in the array must have the same size.

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>