build: force linker to import some dpdk symbols
DPDK PMDs are loaded in a list at program's start using a __constructor
function. Symbols for those functions are not set to a public visibility
so they are ignored during link time of a program using packetgraph.

The strategy with gcc was to use --whole-archive flag so all libdpdk.a
library is included at link time. However this option is not availale
with clang 3.4.2.

The workaround here is to change some symbol visibility to public and to
mark those symbols used in packetgraph so the linker will won't ignore
them.

This has been done for all dpdk's pmd.

Signed-off-by: Jerome Jutteau <jerome.jutteau@outscale.com>