build: autogen version.mk, and fix versioning in Makefile
Until recently packetgraph version was hardcoded in the Makefile,
and we had no way to ghet the version in C.

It's now resolve by adding 4 define in "common.h"
those defines are:
       PG_VERSION_YEAR: year of last DPDK release on which packetgraph is bases
       PG_VERSION_MONTH: month of DPDK release
       PG_VERSION_REVISION: number of revisions on the release
       PG_VERSION: number containing all information above

As we need to got thoses informations for the linker in the makefile,
we are compiling an intermediate program in the configure that ghenerate a
version.mk which is then use in the makefile.

So we are now generating a libpacketgraph.so.19.2.0 instead of
libpacketgraph.so.17.5.0.

Signed-off-by: Matthias Gatto <matthias.gatto@outscale.com>